remove static links to Main
This allows us to access the public properties/methods of kyScreen. While theres nothing inherently wrong with using static references, I feel that we cannot fully utilise the benefits of kyScreen through static references, as you would have to create a static function/property that updates it constantly in the update loop.